Consolidated Integrations Tests
